Changed several matrix partitioning parameters and methods. Enabled a variety of 1d (row, column) partition methods. This is the new system:
MATRIX_APPROACH:
  ROWS    - for row partitioning
  COLUMNS - for column partitioning
  NONZEROS or GENERAL - for general (2d) partitioning

MATRIX_METHOD:
  HYPERGRAPH
  BLOCK
  RANDOM
  more to come...

For 1d partitioning, the MATRIX_METHOD is really just passed on
as the LB_METHOD for the derived partitioning problem. For 2d,
we will need a full native implementation for each method.
Fine-grain hypergraph will be the first one.

A bit confusingly, zdrive will first set LB_METHOD to SPARSE_MATRIX
then LB_METHOD is later redefined as the 1d partitioning method.
SPARSE_MATRIX is not a valid LB_METHOD option in Zoltan,
but is faked as such in zdrive to simplify the driver.