Summary:  Fixing an intermittent MPI error when ML is configured with --enable-ml_timing.
Bug: 3646

Comments: In ML_Operator_Destroy, an global MPI reduction operation was guarded
by an if that checked whether a timer was zero/nonzero.  For a simple 3
processor run, one processor's timer was zero, while the other processors'
timers were nonzero, resulting in only two of the three processors
calling the global reduction.  MPI issued an error.

The fix is to remove that timer from the if check.  Alternatively, the if could
rely on a global sum of the timer.