SUMMARY: Updated Sundance to run the ModelEvaluator example correctly (for Trilinos 8.0 initially)
BUGS: Trilinos bug 3671 (no bug filed for Sundance)

CHECKLIST:

1) Did you add any source or other files needed for building, testing,
   or documentation? No

2) Did you update or change the build system (including adding
   files)? No

3) Did you do a cvs update *very* recently before rebuilding and
   retesting? Yes

4) Did you rerun the test suite on at least the packages that you
   changed (rerunning the entire Trlinos test suite on all affected
   packages is *strongly* recommended)? Yes, all of the tests passed that
   where run on my machine (except the Python tests and the test IDAHeat.tst
   which was inactive).

COMMENTS:

Here I updated the ModelEvaluator example in Sundance.  The main thing that I
had to do was to fix a small defect in Thyra (a one-line change) and then
replace ENABLE_EPETRA_THYRA with ENABLE_THYRA_EPETRA.

While I was at it, I did a few other things:

* SimpleModelEvalTest.cpp:

- I replaced the use of MoochoPack::MoochoSolver with
MoochoPack::MoochoThyraSolver.  The MoochoThyraSolver class is a higher-level
facade that takes care of a lot of the details of hooking up a
Thyra::ModelEvaluator to Moocho to solve an NLP.

- I dropped in the Stratimikos class Thyra::DefaultRealLinearSolverBuilder to
create LOWSFB object.  This removed a lot of boiler-plate code.  I did not put
in the hooks

*SundancePositionalCellPredicate.hpp: Changed the name of CELL_PREDICATE(...)
to CELL_PREDICATE_(...) and removed the ending semicolon.  Then, I create a
new macro CELL_PREDICATE(...) defined as CELL_PREDICATE(...); to support
existing uses.  Now you can write:

  CELL_PREDICATE_(...);

with a semicolon at the end and emacs will indent this correctly.  I hope
people will update their other files to use this new form and then at some
point we can use a sed script to change everyones code from
CELL_PREDICATE_(...); to CELL_PREDICATE(...);.  I would have just done this
change but I know there are lots of Sundance user code that uses this macro
and I did not want to break their usage (without a semicolon at the end).
This gives these users time to upgrade (hopefully).

* AdvectionDiffusion2D.cpp: Checking in code to write matrices out to file.  We
can revert this back at any time but I just wanted to get a version of this
copied so that I can access it later.