SUMMARY: Change of behavior for LinearOpBase and ModelEvaluator::create_W_op() to allow for simpler implementations of ModelEvaluator subclasses and client code.
BUGS: 3676

CHECKLIST:

1) Did you add any source or other files needed for building, testing,
   or documentation? No

2) Did you update or change the build system (other than just adding
   source files)? No

3) Did you do a cvs update *very* recently before rebuilding and
   retesting? Yes, but there are updated CMake files in epetra.

4) If justified, did you type 'make partial-clean' and then rebuild all
   libraries and executables with 'make everything'? Yes

5) Did you rerun the test suite on at least the packages that you
   changed (rerunning the entire Trlinos test suite on all affected
   packages is *strongly* recommended)? Yes, all 777 tests ran passed

COMMENTS:

***
*** BEGIN WARNING !!!!
***

This checkin involves a refactoring where:

   new Thyra::EpetraLinearOp(...)

must be replaced with either:

  Thyra::epetraLinearOp(...)

or

  Thyra::nonconstEpetraLinearOp(...)

depending on if you need a const or non-const object respectively.

I can add back the direct constructor for backward comparibility if desired
(just ask).

***
*** END WARNING !!!!
***


This is the first part of a refactoring that will make it possible to simplify
the creation of block-based and perhaps other types of implicit ME subclasses.

Here are the changes:

(*) Thyra_LinearOpBaseDec.hpp: Here I have changed the specification to allow
for three different initialization states: a) Fully uninitialized, b)
partially initialized, and c) fully initalized.  Previously, there where just
two states.  The new "partially initialized" state allows the range and domain
spaces to be given but not the implementation of the apply(...) function.  I
added non-friend, non-member helper functions to define these states and act
as query functions.

(*) Thyra_ModelEvaluator.hpp: Here I changed the specification of the
post-conditions of ME::create_W_op() to ensure that the range and domain
spaces on the returned W_op object are supported right away.

(*) Thyra_EpetraLinearOp.[hpp,cpp]: Here I added the functions
partialInitialize(...) and setFullyInitialized(...) and updated to allow for
the partial initialization state.  This is the key to allowing the changing
the behavior of the ME::create_W_op() function described above.  Another major
change I made was to remove the direct non-default constructor.  This required
refactoring a bunch of user code to change from using raw 'new
EpetraLinearOp(...)' to using the non-member constructors
'epetraLinearOp(...)' or 'nonconstEpetraLinearOp(...)'.  I can add back the
current constructor if this breaks any user code.

(*) Thyra_EpetraModelEvaluator.[hpp,cpp]: I updated this class to use the new
partial initialization state of Thyra::EpetraLinearOp.  Now when a W_op object
is created, it has range and domain spaces set.  This will allow the
simplification of certain implicit ME subclasses now.  Note that this does
*not* require that EpetraExt::ModelEvaluator::create_W() return fully
initialized Epetra_Operator objects.  This was an important part of the
refactoring in my opinion.  We have not made it any harder to define
EpetraExt::ModelEvaluator subclasses.

(*) Thyra_EpetraTypes.hpp: Changed const char* returns to std::string returns.
No more raw pointers!

(*) ALL OTHER FILES: All of the other files part of this checkin where the
result of replacing raw 'new EpetraLinearOp(...)' with non-member calls
'epetraLinearOp(...) or 'nonconstEpetraLinearOp(...)'.