SUMMARY:   Adding a new EpetraExt transform adapter for Isorropia.
     BUGS:      None.

     CHECKLIST:

     1)   Did you add any source files?  Yes.
     1.a) Did you do a 'cvs add'?  Yes.
     1.b) Did you add them to a Makefile.am file in some way?  Yes.
     1.c) Did you do a new bootstrap?  Yes.

     2)   Did you update or change the build system?  Yes.
     2.a) Did you remember to do a bootstrap and do a 'cvs add' on the
          Makefile.am files and the Makefile.in files?  Yes.
     2.b) [Optional but recommended] Did you blow away the libraries and
          executables and then rebuild?  Yes.

     3) Did you rerun the test suite on at least the packages that you
        changed?  Yes.

     COMMENTS:

     EpetraExt has a graph transformation interface that can take an input graph,
perform some function on it, and then return a new graph.  This interface is used
to apply graph transformations like RCM, AMD, BTF, and graph partitioning.  While
EpetraExt has a graph transformation interface for Zoltan, it would be better to
have an interface to Isorropia, since it is to be the parameter-list driven
interface for graph partitioning in Trilinos.  Since it is not clear the proper
hierarchy of such a interface (I think Isorropia uses some of the other utilities
in EpetraExt), it would be better to place this code in TrilinosCouplings.

     That being said, this is the initial check-in of the EpetraExt-Isorropia transform
interface.  It will be built if --enable-isorropia and --enable-epetraext are enabled.
It also has an example in the example/epetraext directory that shows how to make an
Epetra_LinearProblem and balance it using the EpetraExt_Isorropia_CrsGraph transform.

     All the tests still pass.