Bugs 3771, 3740
Removing the requirement of a Fortran compiler when building AztecOO.  There
are now C equivalents of the two Fortran source files in AzteOO.  After
consulting with Mike H, we decided to leave the Fortran files as the default
for now.

In addition, I cleaned up the problem that we saw with Teuchos and the new
BLAS/LAPACK detection when libcheck is disabled.

I also cleaned up a few obsolete parts of configure.ac and src/Makefile.am and
removed az_fortran_wrap.c.  The Fortran wrappers are obsolete, per Mike H.
This allowed me to remove some obsolete code in other files too.

I recompiled on my machine with fortran enabled and disabled and ran the
runtests make target for both builds as well.

I still need to clean up the two new C files, but I want to get this checked
in because everything is compiling with and without Fortran.