SUMMARY: Fixed const Array iterator usage and then change bug and other small changes.
BUGS: None filed

CHECKLIST:

1) Did you add any source or other files needed for building, testing,
   or documentation? No

2) Did you update or change the build system (other than just adding
   source files)? No

3) Did you do a cvs update *very* recently before rebuilding and
   retesting? Yes

4) If justified, did you type 'make partial-clean' and then rebuild all
   libraries and executables with 'make everything'? No, just Teuchos and
   Intrepid

5) Did you rerun the test suite on at least the packages that you
   changed (rerunning the entire Trlinos test suite on all affected
   packages is *strongly* recommended)? Only ran Teuchos and Intrepid test
   suite in debug and optimized mode (all tests ran passed).

COMMENTS:

Here is what I did:

(*) Teuchos_Array.hpp, Array_test.cpp: I added a test case to expose the
iterator error that was seen in Intrepid.  The issue is that the const
iterator functions create carcp_ while the non-const iterator functions only
create arcp_.  As a result, when resize(...) function tried to change the
array, it was setting arcp_ to null but not carcp_.  I verified that the test
case I added failed before I changed the behavior in Teuchos::Array to fix
this.  I just needed to add logic to set carcp_ = null before checking if the
reference count on vec_ is zero or not.  All tests now pass, as well as the
intrepid tests (both debug and optimized).

(*) Teuchos_ArrayRCP.hpp, ArrayRCP_test.cpp: Added casting operator
arcp_const_cast(...) and added test for this.

(*) Makefile.export.teuchos.in: Added TEUCHOS_DEPS macro.  This will be used
to build in better dependancy tracking to figure out when *.exe files must be
rebuilt in other packages.  This does not currently work yet due to a bug in
gmake?